+++ src/main/java/com/smartech/i2019/adaptiveinterviews/model/Employee.java
@@ -1,5 +1,6 @@
 package com.smartech.i2019.adaptiveinterviews.model;
 
+import lombok.Data;
 import org.hibernate.Hibernate;
 
 import javax.persistence.*;
@@ -9,9 +10,10 @@ import java.util.Iterator;
 import java.util.Set;
 
 @Entity
+//@Data
 @Table(name = "employees")
 public class Employee {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private Department department;
@@ -71,11 +73,11 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
